Avoid recreating channels on every played note
This was a major memory leak, as new channels
would be spun up every time a note was played.
This seems to be a pre-existing bug in SiON, but
perhaps it was somehow patched by a GC or
something.

This commit also adds some clean up code and
optimizations identified during the investigation.

There is one more leak related to playback,
which can be reproduced by playing an MML
melody in the example project. With every play
the number goes up. But it's not an immediate
issue for Bosca Ceoil, so I'm leaving it until a
future date.