Remove unused/WIP simulator code
This code was introduced in the original SiON 12
years ago in
https://github.com/keim/SiON/commit/68f9e5b04d658bf38e128159bbdf737385d35369

With additional work committed in
https://github.com/keim/SiON/commit/0047ca99e71a9953ef2676e6551084490efac6f0

However, the implementation is incomplete to
this day, and the code is not actually used in
any way. It's also not accesssible to external users,
as it needs internal hooks to actually be useful.
Its purpose seems to be to control SiMMLTracks.

I can't judge if the proposed refactoring is a good
idea, but there is no point in keeping this code
in the main branch when it does nothing.

Since time was spent to port it originally, I made
a separate branch from the previous commit to
make the implementation easily discoverable,
should we ever need it.

See branch `wip-simulator`.