CI: Add a codesign step for macOS extension builds
This is not fully tested yet because of Apple issues.
The notarization process can hang in perpetuity, with
no way for the developer to affect it or even to
understand the source of the hiccup.

But the overall flow for the extension codesign should
be correct here. We still need to do the same for the
example project, but let's see if we can get this working first.