Separate operator count and analog-like settings in channel parameters
This is a breaking change, as previously the
analog-like mode was enabled by setting the
operator count to 5 (the real operator count
would then be set to 2).

This should lead to cleaner API and removes
the need to handle special cases. There is also
the added benefit of having validation in
related methods now.