Add synthesized wave data testing
The output is surprisingly consistent in my
testing, so for not we'll be testing without any
deviation. It's possible, though, that the results
may very between different platforms, making it
harder to automate such testing in the future.

But still, this isn't a part of the CI just yet, so no
need to worry too much about that.

Note that the audio server leaks a lot of objects
in the headless mode, and there is no recourse.
So some leaks on exit are expected and should be
ignored.

This also includes some improvements to the printed results.