Further fix circular macro checks
This was completely not operational in the
original synth, but my initial attempt to fix it was
faulty as well. This broke a handful of MML songs
which we showcase in the example app.

Also includes some adjustments to the example
app, namely the volume slider (some MML songs
are way too loud otherwise).