Simplify SLL implementation and remove unnecessary abstractions
This is a part of a rework that should make it
safer to use and properly managed.

Lists and rings can be handled mostly together,
so we don't need two sets of methods. We also
don't need a "pager", at least not as a List of SLL.
It's supposed to be indexed access, but Lists don't
work like that and we can actually implement the
same method in SLL.