Improve validation when parsing MML voices
Original code is very lenient towards mistakes
in the MML data, and will simply loop over the
value to fit it into the valid range. No error is
reported most of the time.

This isn't that great if we want to be able to
create voices from scripting, as mistakes are
hidden. And even MML compilation is worse
for that, so this commit aims to keep the
behavior, but add error reporting at least.

Some parts are further adjusted to better handle
incorrect values.