Adjust multiple parameter of SiOPM voices to be in line with other types
This is probably an artifact of some early code
iteration, but basically unlike other voice/chip
types, SiOPM voices were doing something strange
about their (fine) multiple parameter.

First of all, the code was briefly converting the
value to a double (Number), but that meant
nothing in practice, as both the raw value and the
stored multiple are ints. Then, the exact
assignment was done manually instead of going
through the setter.

After a bit of investigation I figured that neither
of these behaviors is needed anymore. The double
conversion, as I said, was unnecessary, and the
assignment was identical to the setter, except
written in a different way (multiplication instead of
a bit shift).

There was also no defined limit on the multiple
parameter, but everything else expect it to be in
the range from 0 to 15, and the set value is always
treated in this way anyway. So, just like all the
other parameters since yesterday, the multiple of
SiOPM voices is now validated.